pyproject.toml
[project]
name = "notipi"
version = "1.1.0"
description = "A PyPI package to deliver notifications via your own telegram bot"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [ "anyio==4.4.0", "certifi==2024.6.2", "enum34==1.1.10", "h11==0.14.0", "httpcore==1.0.5", "httpx==0.27.0", "idna==3.7", "python-dotenv==1.0.1", "python-telegram-bot==21.2", "sniffio==1.3.1", "ipython==8.25.0", "nest-asyncio==1.6.0"]
authors = [
    { name = "Harshit Sharma", email = "harshit158@gmail.com" }
]

[build-system]
requires = [ "setuptools",]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"notipi" = ["sounds/*.mp3"]

[project.scripts]
noticli = "notipi.notipi:noticli"

[project.urls]
"Homepage" = "https://github.com/harshit158/notipi"
"Source" = "https://github.com/harshit158/notipi"